userinfo added + use name lifted to header
