--- src/client/containers/Profile/index.styles.css
@@ -1,9 +1,0 @@
-.middle {
-  display: flex;
-}
-.left-part {
-  flex: 2;
-}
-.right-part {
-  flex: 1;
-}
